craft level up update

+++ SonOfRobin.Core/Game/CraftStats.cs
@@ -120,20 +120,19 @@ namespace SonOfRobin
             this.DisplaySummary(collectionToShow: this.usedIngredients, header: "Used ingredients");
         }
 
-        public float GetRecipeLevel(Craft.Recipe recipe)
+        public int GetRecipeLevel(Craft.Recipe recipe)
         {
             int craftCount = this.HowManyTimesHasBeenCrafted(recipe);
             float recipeLevel = (float)craftCount / (float)recipe.craftCountToLevelUp;
-
-            return Math.Min(recipeLevel, recipe.maxLevel);
+            return (int)(Math.Min(recipeLevel, recipe.maxLevel));
         }
 
         public bool RecipeJustLevelledUp(Craft.Recipe recipe)
         {
             // repeated calculations, because Math.Min() would cause level up to be reported every time at master level
+
             int craftCount = this.HowManyTimesHasBeenCrafted(recipe);
             float recipeLevel = (float)craftCount / (float)recipe.craftCountToLevelUp;
-
             return recipeLevel > 0 && recipeLevel <= recipe.maxLevel && recipeLevel == Math.Floor(recipeLevel);
         }
